% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcsv.R
\name{exportpkg.csv}
\alias{exportpkg.csv}
\title{Export installed R packages to a CSV file}
\usage{
exportpkg.csv(file)
}
\arguments{
\item{file}{The path of the CSV file.}
}
\description{
This function exports to a CSV file all the currently installed packages
of your installation. You can import it later to another one using \code{installpkg.csv()}.
}
\examples{
exportpkg.csv('exported-packages.csv')
}
\keyword{CSV}
\keyword{backup}
\keyword{export}
\keyword{import}
\keyword{packages}
