% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcsv.R
\name{installpkg.csv}
\alias{installpkg.csv}
\title{Install R packages from a CSV file}
\usage{
installpkg.csv(file)
}
\arguments{
\item{file}{The path of the CSV file.}
}
\description{
This function reads the CSV file, checks which packages are currently installed
and installs the missing packages from the file.
}
\keyword{CSV}
\keyword{backup}
\keyword{export}
\keyword{import}
\keyword{packages}
